Add handling of SIGHUP system signal
This commit adds handling of SIGHUP system signal to AcraTranslator
service. The approach is ported from AcraServer service that already
possess this feature. Basically we need to distinguish how to start
our service: with 'Start' or with 'StartFromFileDescriptor' function.
We control it with a help of 'GracefulEnv' environment boolean variable.