Generate test certificates with SANs
Go 1.15 drops legacy support for CommonName as host name [1] if there
are no Subject Alternate Names in the certificates. The certificates
we use for testing are exactly like that and do not work with Go 1.15.

[1]: https://golang.org/doc/go1.15#commonname

Instead of putting "localhost" into Common Name, add a Subject Alternate
Name with that host name, and use something which does not look like
a hostname in Common Name.

The interface for setting these extension variables is not a paragon of
usability, but OpenSSL is what it is. We need to provide a configuration
file and pass the section names to use. Since those are test certificates,
we can use only bare minimum of fields.