Fix Go 1.15 support with new test certificates (#417)
* Generate test certificates with SANs

Go 1.15 drops legacy support for CommonName as host name [1] if there
are no Subject Alternate Names in the certificates. The certificates
we use for testing are exactly like that and do not work with Go 1.15.

[1]: https://golang.org/doc/go1.15#commonname

Instead of putting "localhost" into Common Name, add a Subject Alternate
Name with that host name, and use something which does not look like
a hostname in Common Name.

The interface for setting these extension variables is not a paragon of
usability, but OpenSSL is what it is. We need to provide a configuration
file and pass the section names to use. Since those are test certificates,
we can use only bare minimum of fields.

* Regenerate certificates and keys

Update the files in the repository with new settings. This will probably
trigger GitHub alarms about private keys being committed into repository
but that's okay.

* Use updated Docker images

The database images we use contain the certificates and keys generated
here, in Acra CE. Switch to updated versions of the images which contain
the freshly generated data.

* Update tested Go versions

Test latest stable version (1.15.x), the previous one (1.14.x), and the
oldest version we support (1.13.x).

* Tweak GOROOT only for general_validation step

This is the only step which cares about the value of this variable.
Everywhere else it's ignored and may be removed (will be done in
a separate PR).