Split TLS certificate and key options as well
Introduce "tls_{client,database}_{cert,key}" options too, allowing
AcraServer to use different certifcates for AcraConnector and database
connections.

With this, TLS configurataion is now completely separated. It is
possible to configure AcraConnector and database TLS connections
separately if necessary, but there is also a common option set
if that is more convenient.