Unique session IDs on ClientSession's logs
Make sure that each session gets a unique "session_id" field in logs.
This makes is really trivial to filter logs for activity of a particular
session.

Don't be too smart with GUIDs or something, keep the field as short as
possible. We only need it to be unique for a relatively short timeframe.