Use ClientSession in session proxy constructors
Instead of passing just two connections and an unrelated context, pass
a ClientSession instance over to the proxy. Introduce an interface which
allows access to interesting fields of AcraServer's ClientSession.

Since the session is now actually used to retrieve the context, we can't
just pass "nil" values in tests. Introduce stubs to use instead.