Prepared statement registry for PostgreSQL
Introduce PgPreparedStatementRegistry for PostgreSQL-specific prepared
statements. This is pretty basic implementation. It *may* turn out that
we don't really need to have it PostgreSQL-specific, it might get used
by MySQL as well as Acra EE too. However, for now, keep it inside the
PostgreSQL module.

Note that PgProxy keeps around the downcasted statement registry so that
it's able to use extended API, not constrained by the "base" interface.
If there is already a registry in the session, it also checks that is
has the correct type.