Extract ClientSession interface (#420)
* Remove unused fields from ClientSession

Only ClientCommandsSession which inherits from ClientSession uses
keystore. ClientSession does not need to access the keystore, so
remove it from there.

Similarly, only ClientCommandsSession makes use of a backreference to
SServer instance, there is no need for it in ClientSession.

* Refactor ClientCommandSession construction

There is no real need for ClientCommandSession to inherit from
ClientSession, there is no shared behavior. Instead of using it
for shared fields, just introduce the same fields. Instead of
constructing a new context, reuse the one provided by the caller.
Also, avoid using public fields, just pass the server reference
to the constructor and store it there.

* Move HandleClientConnection to SServer

Instead of being a part of ClientSession, this actually belongs to
AcraServer which decides how to use a session. This commits makes
a crude move of the method implementation, later ones will clean it
up a bit and avoid direct field access.

* Export ClientSession's Close() method

Since it's used by SServer, it's not nice to keep it unexported. Close()
is a conventional Go method to close things, it's not wrong to export it.

* Use ClientSession's logger via accessor

Instead of using it directly, introduce a Logger() accessor. This will
also come handy later for decryption proxies.

Additionally, make the log lines a bit more concise by caching the
logger in a local variable (with the same name as imported logrus).

* Introduce other ClientSession accessors

Instead of accessing other fields directly, use accessor methods.
Those will be useful later for decryption proxies too.

* Unique session IDs on ClientSession's logs

Make sure that each session gets a unique "session_id" field in logs.
This makes is really trivial to filter logs for activity of a particular
session.

Don't be too smart with GUIDs or something, keep the field as short as
possible. We only need it to be unique for a relatively short timeframe.

* Use ClientSession in session proxy constructors

Instead of passing just two connections and an unrelated context, pass
a ClientSession instance over to the proxy. Introduce an interface which
allows access to interesting fields of AcraServer's ClientSession.

Since the session is now actually used to retrieve the context, we can't
just pass "nil" values in tests. Introduce stubs to use instead.

* Simplify "HandleClientSession" a bit more

Since this is a method of SServer now there is no need to pass proxy
factory as an argument. It is accessible as a field.