Add TODO to "Add()" about updates
As we currently don't expect the implementation to accurately track
prepared statement removals, allow the Add() calls to update the
contents of prepared statements and cursors. However, the actual
PostgreSQL protocol does not allow that and requires the clients
to send Close messages before redefining named entities. However,
there is an exception for unnamed statements and portals.

Expect that Add() will refuse updates in the future and do not return
an additional boolean result. Leave a TODO in the code so that we don't
forget to forbid updates for named things.