Removal of statements and cursors
Add methods to delete prepared statements and cursors from the registry.
They are a bit involved since removing a prepared statement implicitly
removes all cursors associated with it.

Also, note that we do not treat removal of nonexistent statements and
cursors as an error. This mimicks PostgreSQL protocol which allows
frontends to send Close messages with arbitrary names.