Store PgProtocolState in ClientSession
The protocol state is fundamentally associated with the session so it
makes sense to keep one there. However, we need to keep it opaque as
MySQL and PostgreSQL protocols are very different. For now, let's use
just "interface{}" as a placeholder. Maybe later we will be able to
extract some more specific interface out of the implementations.