Clear pending query on ReadyForQuery
PgProtocolState keeps user query in memory until the database issues
a response. The query text is sensitive information, especially when
transparent encryption is in use. We should remove it from memory as
soon as it possible. The database sends ReadyForQuery when it has
completed processing of the previous query, this is the point when we
know that the query is no longer needed and can be freed.