Keep cursors of prepared statements in them directly
Actually, we *can* maintain lists of cursors of prepared statements
in them directly. This makes it easier to iterate over cursors of
particular prepared statement -- for example, when we need to remove
it from the registry.

Simultaneously, we keep a mapping of all cursors by their name since
that's how they are going to be requested for data.