Prepared statement and cursor registries (#422)
* PreparedStatement and PreparedStatementRegistry interfaces

Introduce interfaces for accessing the prepared statement registry of
a session as well as to initialize it.

The prepared statement registry -- as it name implies -- will be used to
keep track of prepared statements created within a session.

* Prepared statement registry for PostgreSQL

Introduce PgPreparedStatementRegistry for PostgreSQL-specific prepared
statements. This is pretty basic implementation. It *may* turn out that
we don't really need to have it PostgreSQL-specific, it might get used
by MySQL as well as Acra EE too. However, for now, keep it inside the
PostgreSQL module.

Note that PgProxy keeps around the downcasted statement registry so that
it's able to use extended API, not constrained by the "base" interface.
If there is already a registry in the session, it also checks that is
has the correct type.

* Cursor and CursorRegistry interfaces

Just like with prepared statements, add a couple of interfaces for
access to the cursors used to iterate over prepared statements.

* Portal registry for PostgreSQL

Introduce PgPortalRegistry to keep track of PostgreSQL portals -- that's
how cursors are called internally in the protocol. It is mostly similar
to the prepared statement registry in the implementation.

* Simplify initialization of registries

Don't do any fancy downcasts or anything. Try to keep usage to the
interface provided by "base" module. This will allow us to track whether
it is actually necessary to have database specific registries or we can
go with a single common utility implementation.

* Add TODO to "Add()" about updates

As we currently don't expect the implementation to accurately track
prepared statement removals, allow the Add() calls to update the
contents of prepared statements and cursors. However, the actual
PostgreSQL protocol does not allow that and requires the clients
to send Close messages before redefining named entities. However,
there is an exception for unnamed statements and portals.

Expect that Add() will refuse updates in the future and do not return
an additional boolean result. Leave a TODO in the code so that we don't
forget to forbid updates for named things.

* Merge statement and cursor registries

After further research into how to implement statement removal I have
found that they are tied to cursors more tightly than expected. Thus
I believe that it makes more sense to provide a unified "prepared
statement registry" which includes cursor support too. In fact, cursors
don't really make sense alone, without associated prepared statements.

* Use the name "cursor" in most cases

It causes countless troubles when writing more 'generic' code to call
cursors portals. Use the more common term where possible, outside of
clearly PostgreSQL-specific contexts.

* Keep track of prepared statement cursors

In addition to prepared statement and cursor mappings, for each prepared
statement keep around a map of cursors associated with it. This will
allow to drop all of them from registry when removing the statement.

* Removal of statements and cursors

Add methods to delete prepared statements and cursors from the registry.
They are a bit involved since removing a prepared statement implicitly
removes all cursors associated with it.

Also, note that we do not treat removal of nonexistent statements and
cursors as an error. This mimicks PostgreSQL protocol which allows
frontends to send Close messages with arbitrary names.

* Unit tests for prepared statement registry

This object has become complex enough to warrant some unit tests as
I'm not sure anymore that it will behave correctly and want additional
guarantees for that.

* Keep cursors of prepared statements in them directly

Actually, we *can* maintain lists of cursors of prepared statements
in them directly. This makes it easier to iterate over cursors of
particular prepared statement -- for example, when we need to remove
it from the registry.

Simultaneously, we keep a mapping of all cursors by their name since
that's how they are going to be requested for data.