Log and keep processed prepared statement data
Now we're interested in other parts of the Parse packet, not only the
query. Change GetParseQuery() to GetParseData() which returns the entire
ParsePacket data, not just the query string.

Add an accessor to prepared statement name. ParsePacket has a peculiar
layout where the name and query strings are kept *with* the trailing
null bytes. Cut those off.