Parse Bind and BindComplete packets
Bind packets carry query parameters for prepared statements issued with
Parse packets. Acra may need to apply transparent encryption and other
transformations to parameter values. We also need the portal name from
the packet to register the corresponding cursor, so that when the data
of the response comes we know how to interpret it, how to match it to
table columns for decryption, etc.

The content of the Bind packet is relatively easy to parse, just read
the linked PostgreSQL spec if you need a reference. Currently we're
interested only in the statement and portal names, other accessors will
be added later as needed.