Process inbound Bind packets
Handle the Bind packets sent by the client, remember them until we get
a response from the database, confirming the portal creation.

This is also the place where we should apply OnQuery observers to the
bound parameter data, but we will do it at some later point. Right now
we lack the capabilities for that.