Parse Execute packets
Execute packets carry the name of the portal to query (and some
auxiliary data). Acra is not really interested in processing it
before it reaches the database, but we'll need it later when the data
responses start coming. This allows Acra to identify the query for which
the responses are returned.