TODO: zeroize packet data
Intercepted packets contain sensitive data: SQL queries as issued by the
user. While they are kept in a relatively safe place -- AcraServer process
memory -- this is still sensitive data which should be cleaned out when
we don't need it anymore.

Make sure that packet content gets zeroized (in the future). I don't have
a nice idea on how to do this right now, but this should be done.