Observing bound values of prepared statements
Add OnBind() method to the QueryObserver interface used for transparent
encryption processing. This will allow to implement transparent
encryption for prepared statement parameters. The parameters refer to
the query but the query cannot be changed, only the parameters may be.

Parameter values are encapculated in the BoundValue interface, similar
to OnQueryObject interface for the OnQuery() method.

We also add skeleton implementations of OnBind() to all exisiting
QueryObservers: the multi-observer helper and QueryDataEncryptor.
Here we are interested in INSERT and UPDATE queries where transparent
encryption is going to happen.