Return original query on error in ArrayQueryObserverableManager
Refactor the code so that it keeps track of the current query when
applying a chain of observers (so that if one of them changes the
query, the following observers will see an updated query). However,
if any observer fails to process a query, return the *original*,
unmodified query -- as indicated by the "false" value -- instead of
returning the currently processed one.