IDs for QueryDataEncryptors (#427)
* Add ID() method to QueryObserver interface

This is similar to DecryptionSubscriber used for query decryption, but
for the encryption code path. It will be used in the same manner: for
improved logging in multi-observer utilities.

* Do not log query in ArrayQueryObserverableManager

The query can be quite big, and it's already logged in an appropriate
manner in session proxies, with proper sanitation. There is no need
to duplicate this log here.

* Return original query on error in ArrayQueryObserverableManager

Refactor the code so that it keeps track of the current query when
applying a chain of observers (so that if one of them changes the
query, the following observers will see an updated query). However,
if any observer fails to process a query, return the *original*,
unmodified query -- as indicated by the "false" value -- instead of
returning the currently processed one.

* Log faulty OnQueryObserver in ArrayQueryObserverableManager

If one of the observers fails then log this failure with attached
observer ID, like ColumnDecryptionObserver does for decryption.