PgProtocolState handler (#424)
* PgProtocolState handler

Introduce a separate object to take care of PostgreSQL protocol packets.
With extended protocol we will need to track both requests and responses
before we are able to process a query, so we need a place to store this
information temporarily.

* Store PgProtocolState in ClientSession

The protocol state is fundamentally associated with the session so it
makes sense to keep one there. However, we need to keep it opaque as
MySQL and PostgreSQL protocols are very different. For now, let's use
just "interface{}" as a placeholder. Maybe later we will be able to
extract some more specific interface out of the implementations.

* Forward client packets through PgProtocolState

Move in-place handling of packets with queries to PgProtocolState. Now
it will inspect all packets, not only queries, and will decide whether
a packet is interesting, or it should be forwarded to the database
immediately.

* Forward database packets through PgProtocolState

Similarly to client packets, let PgProtocolState observe database
responses as well. Currently there is not much to see, but later
it will be more involved with processing extended protocol and
extracting enhanced type information from responses.

* Clear pending query on ReadyForQuery

PgProtocolState keeps user query in memory until the database issues
a response. The query text is sensitive information, especially when
transparent encryption is in use. We should remove it from memory as
soon as it possible. The database sends ReadyForQuery when it has
completed processing of the previous query, this is the point when we
know that the query is no longer needed and can be freed.

* Fix minor typos

* Split client packet processing into methods

It's quite annoying to duplicate packet forwarding code when packet
handling takes a shortcut. Extract some parts of ProxyClientConnection
into separate utility methods, making it more visible that we always
forward the processed packet to the database, unless an error occurs
or if AcraCensor denies the packet.

* Split database packet processing into methods

Similarly, split out a part of ProxyDatabaseConnection into a separate
method to avoid duplicating the code responsible for forwarding the
packet to the client, stopping the timer span, and error handling.

* Don't special-case termination packet in PgProtocolState

Since it is already handled by PgProxy, we don't need a special packet
type for it.

* Improve packet type naming

* Split processing into even more methods

Extract query data processing into another helper method to call from
generic client/database packet processor. This will allow us to reuse
this code later for the multitude of packets which contain query data.
Right now it's only Query and Parse, but later Bind and Execute will
come into the mix and there is likely a shared part for them.

* Tweak PgProtocolState interface for packet types

Instead of returning a PacketType from handlers, remember the last
packet type we've seen so far. It can be retrieved later multiple
times, and it's a good indication of what kind of packet we expect
to come next.