Handling Bind packets
Now we're ready to handle the Bind packets. This is performed similar
to Parse packets, but now we need to retrieve an extra context: the
prepared statement for which the bind is performed. After we have it,
we can inspect the prepared statement values.

Note that in case of client error *we* have to deal with an invalid Bind
packets, such as referring to a nonexistent prepared statement. In such
cases we simply let the packets pass to the database unchanged. Then the
database is likely to return an error response and this won't go anywhere.

Also note that the parameter count after processing might be different
from the original input provided by the client. This is why OnBind() has
the interface is has: instead of in-place modifications of values, it
returns a new list of possibly modified values. Right now we won't be
using this capability, but in the future this should allow us to handle
queries like this:

    INSERT INTO table(col1, col2) VALUES ($1, $1);

by encrypting "col1" and "col2" values separately, into two different
values. For that we may need to rewrite the query into

    INSERT INTO table(col1, col2) VALUES ($1, $2);

and when the client sends a Bind with one parameter, rewrite that too
into a Bind with two parameters.