Marshal and send updated Bind packets
After we have computed updated Bind values, they need to be marshalled
into the Bind packet to be sent to the database. Introduce MarshalInto()
method to write Bind packet data into the packet payload buffer. It may
be possible that the new data cannot be serialized into a packet. In
that case, emit an error and send the original packet to the database.