Transparent encryption: INSERT parameters
Implement transparent encryption of INSERT parameters. For example,
if the user prepares the following query:

    INSERT INTO table(col1, col2, col3) VALUES ($1, $2, 'user');

then we should encrypt the values of $1 and $2 parameters. However, we
should not and cannot encrypt the 'user' string when processing bound
parameters. That value should have been encrypted when the statement
had been prepared.

In order to perform transparent encryption we need to map the parameter
placeholders ($1, $2...) onto the table columns. That way we can consult
the database schema from AcraServer configuration file and see whether
a column needs to be encrypted and in what way.

We gather this information by analyzing the query. Note that not all
queries are supported for transparent encryption. Only those that are
simple enough for Acra to analyze will be encrypted. Other queries will
be passed to the database unchanged. Or rather, the parameters of such
queries will not be encrypted.