Observe extended query protocol (#426)
* Log and keep processed prepared statement data

Now we're interested in other parts of the Parse packet, not only the
query. Change GetParseQuery() to GetParseData() which returns the entire
ParsePacket data, not just the query string.

Add an accessor to prepared statement name. ParsePacket has a peculiar
layout where the name and query strings are kept *with* the trailing
null bytes. Cut those off.

* Differentiate between simple and extended queries

Make it possible to tell whether we received a Query packet or Parse
packet before, because only for Parse packets you can get their data.

Now that we have this, log the name of the prepared statement which is
going to be sent out to the database.

* Register confirmed prepared statements

Parse and react to ParseComplete packets from the database, confirming
the prepared statements from the previous Parse packet sent by the
client. Once we get ParseComplete, we know that the database is ready to
receive further requests for the prepared statement in question.

* Parse Bind and BindComplete packets

Bind packets carry query parameters for prepared statements issued with
Parse packets. Acra may need to apply transparent encryption and other
transformations to parameter values. We also need the portal name from
the packet to register the corresponding cursor, so that when the data
of the response comes we know how to interpret it, how to match it to
table columns for decryption, etc.

The content of the Bind packet is relatively easy to parse, just read
the linked PostgreSQL spec if you need a reference. Currently we're
interested only in the statement and portal names, other accessors will
be added later as needed.

* Process inbound Bind packets

Handle the Bind packets sent by the client, remember them until we get
a response from the database, confirming the portal creation.

This is also the place where we should apply OnQuery observers to the
bound parameter data, but we will do it at some later point. Right now
we lack the capabilities for that.

* Register confirmed portals

React to BindComplete packets from the database, confirming cursors
requested by previous Bind packets sent by the client. Once we get
BindComplete, we know that a database is ready to receive queries
against the requested portal.

* Copy data buffers from PacketHandler

PacketHandler is a stateful packet parser. The data in descriptionBuf
is valid only until the next packet comes. Since we'd like to save
the resulting {Parse,Bind,Execute}Packet objects, we need to make
a persistent copy of the packet buffer which will not be changed.

* Parse Execute packets

Execute packets carry the name of the portal to query (and some
auxiliary data). Acra is not really interested in processing it
before it reaches the database, but we'll need it later when the data
responses start coming. This allows Acra to identify the query for which
the responses are returned.

* Remember inbound Execute packets

So, we don't need to process Execute packets so there is no new packet
type, but we still keep the packet around so that we can query it later
to find out what cursor the data comes from.

* Introduce ZeroizeBytes utility function

Provide a generic way to zeroize byte slices which are not keys,
just to avoid awkwardly using ZeroizeSymmetricKey() for that.

* Zeroize packet data when possible

Intercepted packets contain sensitive data: SQL queries as issued by the
user. While they are kept in a relatively safe place -- AcraServer process
memory -- this is still sensitive data which should be cleaned out when
we don't need it anymore.

Add Zeroize() method to each packet data type which wipes anything
sensitive from memory. Make sure to call it more setting the values
to nil so that sensitive parameter data does not stay in memory longer
than necessary until the text GC cycle.

Unfortunately, we can't zeroize the OnQuery object since it stores query
data as a string (or as parsed SQL object). It's difficult to access the
underlying byte slice there. We leave this object untouched.

* Fix ineffassign warnings

Since Bind packet ends with the result format array, we don't need the
data slice anymore after that. It should be empty. (We don't check.)

* Optimize ZeroizeBytes

Avoid using FillSlice(0, ...) for zeroization since Go compiler fails to
inline such call and it has much worse performance than inlined loop.

* Check for truncated Execute packets

Don't just blindly call Uint32(), make sure that Execute packet has
enough data with the requested row count.