Remove FillSlice and use fast zeroization (#428)
* Use zeroization utilities instead of FillSlice

Avoid using "FillSlice(0, ...)" for zeroization since this call is
typically not inlined by Go compiler and ends up much slower.
Instead, call dedicated zeroization utilities that we now have.

Also, don't zeroize *encrypted* data in ReadData() implementations after
it has been decrypted. The decrypted data is much more sensitive and we
leave it around.

Drop ZeroizeKeys() from keystore/v2 module since it does not seem to be
used anywhere anymore.

* Lower ineffassign error limit

Recent updates have removed two instances of this error so we can lower
the expected count accordingly.

* Replace last usage of FillSlice in the code base

This is last place where this utility function is used, and this is not
zeroization. Keeping a utility function just for this call site is dumb,
so just inline it.

* Drop utils.FillSlice()

Finally, once all uses of this utility have been removed, drop the
function from the code once and for all.