Improve placeholder parsing and handling
Move placeholder handling into a utility method, and make sure that we
check that placeholders are all valid and consistent. That is, they
refer to existing values, and each placholder uniquely identifies
a column. If that's not the case, the query will be skipped and not
encrypted as it's considered 'too complex'.

Also, leave a placeholder (get it?) for MySQL implementation later.