Transparent encryption of prepared statement parameters (#429)
* Observing bound values of prepared statements

Add OnBind() method to the QueryObserver interface used for transparent
encryption processing. This will allow to implement transparent
encryption for prepared statement parameters. The parameters refer to
the query but the query cannot be changed, only the parameters may be.

Parameter values are encapculated in the BoundValue interface, similar
to OnQueryObject interface for the OnQuery() method.

We also add skeleton implementations of OnBind() to all exisiting
QueryObservers: the multi-observer helper and QueryDataEncryptor.
Here we are interested in INSERT and UPDATE queries where transparent
encryption is going to happen.

* Parameter value accessors for BindPacket

Add GetParameters() and SetParameters() accessors in BindPacket,
operating on the parameter values. The values are stored in the packet
in a particular manner which must be respected. See PostgreSQL docs
for more details on the layout.

* Handling Bind packets

Now we're ready to handle the Bind packets. This is performed similar
to Parse packets, but now we need to retrieve an extra context: the
prepared statement for which the bind is performed. After we have it,
we can inspect the prepared statement values.

Note that in case of client error *we* have to deal with an invalid Bind
packets, such as referring to a nonexistent prepared statement. In such
cases we simply let the packets pass to the database unchanged. Then the
database is likely to return an error response and this won't go anywhere.

Also note that the parameter count after processing might be different
from the original input provided by the client. This is why OnBind() has
the interface is has: instead of in-place modifications of values, it
returns a new list of possibly modified values. Right now we won't be
using this capability, but in the future this should allow us to handle
queries like this:

    INSERT INTO table(col1, col2) VALUES ($1, $1);

by encrypting "col1" and "col2" values separately, into two different
values. For that we may need to rewrite the query into

    INSERT INTO table(col1, col2) VALUES ($1, $2);

and when the client sends a Bind with one parameter, rewrite that too
into a Bind with two parameters.

* Marshal and send updated Bind packets

After we have computed updated Bind values, they need to be marshalled
into the Bind packet to be sent to the database. Introduce MarshalInto()
method to write Bind packet data into the packet payload buffer. It may
be possible that the new data cannot be serialized into a packet. In
that case, emit an error and send the original packet to the database.

* Test transparent encryption of bound values

Add a new test to verify transparent encryption of bound values in
prepared statements. Inherit from existing transparent encryption test
to keep the coverage. However, now make sure that prepared statements
are used to insert and update values in the database instead of using
SQLAlchemy for that.

Unfortunately, I could not find a way to get SQLAlchemy to compile
queries into proper prepared statement syntax, therefore the tests
'compile' the necessary SQL manually.

The tests currently cover only 'binary' prepared statements. That is,
using PostgreSQL protocol features. This is the most common way to
use prepared statements across the database drivers. In particular,
in Java drivers.

* Transparent encryption: INSERT parameters

Implement transparent encryption of INSERT parameters. For example,
if the user prepares the following query:

    INSERT INTO table(col1, col2, col3) VALUES ($1, $2, 'user');

then we should encrypt the values of $1 and $2 parameters. However, we
should not and cannot encrypt the 'user' string when processing bound
parameters. That value should have been encrypted when the statement
had been prepared.

In order to perform transparent encryption we need to map the parameter
placeholders ($1, $2...) onto the table columns. That way we can consult
the database schema from AcraServer configuration file and see whether
a column needs to be encrypted and in what way.

We gather this information by analyzing the query. Note that not all
queries are supported for transparent encryption. Only those that are
simple enough for Acra to analyze will be encrypted. Other queries will
be passed to the database unchanged. Or rather, the parameters of such
queries will not be encrypted.

* Transparent encryption: UPDATE parameters

Similar to INSERT queries, add support for transparent encryption of
UPDATE queries. In the same vein, analyze the query to find out what
columns should be encrypted and what placeholder values match them.

Just like with INSERT, AcraServer supports only a simple enough subset
of queries which are direct updates of the table:

    UPDATE table SET col1 = $1, col2 = $2, col3 = 'fixed value';

We don't support more complex ones. This might come later, or might not
be possible at all (e.g., queries updating table from another table).

* Rename utility methods to avoid confusion
* Remove unneeded utilities to improve readability
* Check for short parameter format arrays
* Use standard limit constants from "math"
* Extract parameter encryption into helper function
* Check for empty columns earlier
* Avoid allocations if no values are going to be encrypted
* Explicit warning about text format not supported
* Fix typos
* Improve placeholder parsing and handling

Move placeholder handling into a utility method, and make sure that we
check that placeholders are all valid and consistent. That is, they
refer to existing values, and each placholder uniquely identifies
a column. If that's not the case, the query will be skipped and not
encrypted as it's considered 'too complex'.

Also, leave a placeholder (get it?) for MySQL implementation later.

* Support multiple VALUES clauses after all

We can encrypt multiple VALUES clauses in INSERT statements, but only if
their usage of placeholders is consistent. That is, if none of the rows
uses previously used placeholder with a different column.

* Log a warning for unsupported ON DUPLICATE key clauses
* Fix encryption error handling

Treat all encryption errors as fatal and just let the packet pass
unchanged. The only exception is if we detect that a value is already
encrypted, in which case it is unchanged and we continue.

* Rename bound value "encoding" => "format"

This entity is called "format" by PostgreSQL and this word seems to be
more commonly used, as in "binary format" and "text format". Rename it
from confusing encoding.

* Base mix-in for testing with PostgreSQL binary protocol

Introduce a shared mix-in for tests with "asyncpg" query executor.
Put there some common code which sets up two executors connected
to AcraConnectors with "keypair1" and "keypair2" clients, commonly
used in tests.

Replace all ad-hoc setup of AsyncpgExecutor with this shared class.
It can also be used by Acra EE tests if necessary.

* Query compiler for PostgreSQL prepared statements

Provide a utility compileQuery() method to compile SQLAlchemy query
objects into SQL with proper PostgreSQL prepared statement placeholders
usable by asyncpg executor.

* Control literal bind compilation in compileQuery

For the benefit of some tests, allow literal compilation of queries
by passing "literal_bind=True" to compileQuery.

* Rename "BaseBinaryPostgreSQLTestCase"

Since it's actually a base class, not a pure mix-in.