Text format support for "asyncpg" executor
Teach AsyncpgExecutor to use text format for some types of prepared
statement parameters (integers and floats). This can be controlled
with a new field "format" in the ConnectionArgs. Set it to "text"
to request a text format. The default is to use "binary" format.

BaseBinaryPostgreSQLTestCase supports a class variable FORMAT which
is the format passed to AsyncpgExecutors it creates. By default it
uses binary format, but you can set it to "text" per test case.

Acra CE does not use this functionality, but Acra EE does.