Basic support of text format for parameters
It turns out that clients don't really use text formats for data types
which are interesting to Acra CE's transparent encryption. It works only
with byte arrays which are typically using binary format.

Text format may be occasionally used for types interesting to Acra EE:
integers and floats. However, Acra EE's encryptors *expect* such data in
text format in the first place.

Therefore, for text format of parameters we need to pass them as is.
It's the binary format which requires more attention. If the parameter
is integer -- it still must be encoded as text for the encryptor, even
if the parameter is binary.  We'll need to know what type the parameter
has and then possibly decode and reencode it after encryption. Right now
that's not possible so we continue passing binary data as is.

Passing through text parameters lets Acra EE's encryptors work correctly
when text formats are used for interesting data types. Let's add this
part at first and then get back to binary parameter recoding.